toolachieving/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
faiss-cpu
numpy
pypdf
python-docx
pandas
openpyxl